use Entry struct from datalogtool
